Refactor how resources are setup and verified
