+++ x-pack/plugins/profiling/server/lib/setup/README.md
