--- x-pack/plugins/profiling/server/lib/setup/steps/get_cluster_settings_step.ts
@@ -1,32 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { ProfilingSetupStep, ProfilingSetupStepFactoryOptions } from '../types';
-
-const MAX_BUCKETS = 150000;
-
-export function getClusterSettingsStep({
-  client,
-}: ProfilingSetupStepFactoryOptions): ProfilingSetupStep {
-  return {
-    name: 'cluster_settings',
-    hasCompleted: async () => {
-      const settings = await client.getEsClient().cluster.getSettings({});
-      const maxBuckets = settings.persistent.search?.max_buckets;
-      return maxBuckets === MAX_BUCKETS.toString();
-    },
-    init: async () => {
-      await client.getEsClient().cluster.putSettings({
-        persistent: {
-          search: {
-            max_buckets: MAX_BUCKETS,
-          },
-        },
-      });
-    },
-  };
-}
