+++ x-pack/plugins/profiling/server/lib/setup/types.ts
@@ -10,13 +10,7 @@ import { PackagePolicyClient } from '@kbn/fleet-plugin/server';
 import { Logger } from '@kbn/logging';
 import { ProfilingESClient } from '../../utils/create_profiling_es_client';
 
-export interface ProfilingSetupStep {
-  name: string;
-  init: () => Promise<void>;
-  hasCompleted: () => Promise<boolean>;
-}
-
-export interface ProfilingSetupStepFactoryOptions {
+export interface ProfilingSetupOptions {
   client: ProfilingESClient;
   soClient: SavedObjectsClientContract;
   packagePolicyClient: PackagePolicyClient;
